+++ packages/map-interface/src/location-details/fossil-collections/index.ts
@@ -1,7 +1,10 @@
 import h from "@macrostrat/hyper";
 import PBDBCollections from "./collections";
+import { useDeprecationWarning } from "@macrostrat/ui-components";
 
 export function FossilCollections(props) {
+  useDeprecationWarning("FossilCollections");
+
   const { data, expanded = false } = props;
 
   if (!data || data.length <= 0) {