+++ packages/mapbox-react/src/easing-legacy.ts
@@ -1,0 +1,84 @@
+import { useMapRef } from "./context.ts";
+import { useEffect, useRef } from "react";
+import type { LngLatBoundsLike, PaddingOptions } from "mapbox-gl";
+import { FlyToOptions } from "./easing.ts";
+
+/**
+ * Ease the map to a center position with optional padding.
+ * @deprecated Use useMapEaseTo instead
+ */
+export function useMapEaseToCenter(position, padding) {
+  const mapRef = useMapRef();
+
+  const prevPosition = useRef<any>(null);
+  const prevPadding = useRef<any>(null);
+  // Handle map position easing (for both map padding and markers)
+  useEffect(() => {
+    console.warn(
+      "Using deprecated function useMapEaseToCenter, consider using useMapEaseTo instead",
+    );
+    const map = mapRef.current;
+    if (map == null) return;
+    let opts: FlyToOptions = null;
+    if (position != prevPosition.current) {
+      opts ??= {};
+      opts.center = position;
+    }
+    if (padding != prevPadding.current) {
+      opts ??= {};
+      opts.padding = padding;
+    }
+    if (opts == null) return;
+    if (prevPadding.current == null) {
+      opts.duration = 0;
+    } else {
+      opts.duration = 800;
+    }
+    map.flyTo(opts);
+    map.once("moveend", () => {
+      /* Waiting until moveend to update the refs allows us to
+      batch overlapping movements together, which increases UI
+      smoothness when, e.g., flying to new panels */
+      prevPosition.current = position;
+      prevPadding.current = padding;
+    });
+  }, [position, padding, mapRef.current]);
+}
+
+/**
+ * Ease the map to a set of bounds, with optional padding.
+ * @deprecated Use useMapEaseTo instead
+ */
+export function useMapEaseToBounds(
+  bounds: LngLatBoundsLike,
+  padding: PaddingOptions | number = 0,
+) {
+  const mapRef = useMapRef();
+
+  const prevPosition = useRef<any>(null);
+  const prevPadding = useRef<any>(null);
+  // Handle map position easing (for both map padding and markers)
+  useEffect(() => {
+    console.warn(
+      "Using deprecated function useMapEaseToBounds, consider using useMapEaseTo instead",
+    );
+    const map = mapRef.current;
+    if (map == null) return;
+    if (bounds == prevPosition.current || padding == prevPadding.current) {
+      return;
+    }
+    let opts: FlyToOptions = {
+      padding,
+      duration: prevPadding.current == null ? 0 : 800,
+    };
+
+    map.fitBounds(bounds, opts);
+    map.once("moveend", () => {
+      /* Waiting until moveend to update the refs allows us to
+      batch overlapping movements together, which increases UI
+      smoothness when, e.g., flying to new panels */
+      prevPosition.current = bounds;
+      prevPadding.current = padding;
+    });
+  }, [bounds, padding, mapRef.current]);
+}